internal/database: merge repos_db_test.go into repos_test.go
This merges the two files, as proposed in Slack. I changed *some*
obvious bits in the file to make things more consistent, but I'll need
another pass.
